donation

+++ app/Http/Controllers/PaymentController.php
@@ -14,15 +14,22 @@ class PaymentController extends Controller
 {
     public function handlePayment(Request $request)
     {
-        $cartId = $request->cart_id;
-        $cart = Cart::with(['detail' => function ($query) {
-                $query->with(['product' => function ($q) {
-                    $q->with('images');
-                }])->with(['color','size']);
-            }])
-            ->where('id',$cartId)
-            ->first();
-        $amount = round((floatval($cart->total_amount)+floatval($cart->delivery_carge)),2);
+        if($request->type == 'donation'){
+            $amount = round((floatval($request->amount)),2);
+            $return_url = route('success.payment');
+        }else{
+            $cartId = $request->cart_id;
+            $cart = Cart::with(['detail' => function ($query) {
+                    $query->with(['product' => function ($q) {
+                        $q->with('images');
+                    }])->with(['color','size']);
+                }])
+                ->where('id',$cartId)
+                ->first();
+            $amount = round((floatval($cart->total_amount)+floatval($cart->delivery_carge)),2);
+
+            $return_url = route('success.payment').'?cart_id='.$cartId;
+        }
 
         $provider = new PayPalClient;
         $provider->setApiCredentials(config('paypal'));
@@ -31,7 +38,7 @@ class PaymentController extends Controller
             "intent" => "CAPTURE",
             "application_context" => [
                 // "return_url" => route('success.payment'),
-                "return_url" => route('success.payment').'?cart_id='.$cartId,
+                "return_url" => $return_url,
                 "cancel_url" => route('cancel.payment'),
             ],
             "purchase_units" => [
@@ -76,46 +83,48 @@ class PaymentController extends Controller
         if (isset($response['status']) && $response['status'] == 'COMPLETED') {
 
             $cartId = $request->cart_id;
-            $cart = Cart::with(['detail' => function ($query) {
-                    $query->with(['product' => function ($q) {
-                        $q->with('images');
-                    }])->with(['color','size']);
-                }])
-                ->where('id',$cartId)
-                ->first();
-            $amount = round((floatval($cart->total_amount)+floatval($cart->delivery_carge)),2);
+            if($cartId){
+                $cart = Cart::with(['detail' => function ($query) {
+                        $query->with(['product' => function ($q) {
+                            $q->with('images');
+                        }])->with(['color','size']);
+                    }])
+                    ->where('id',$cartId)
+                    ->first();
+                $amount = round((floatval($cart->total_amount)+floatval($cart->delivery_carge)),2);
 
-            $transaction = Transaction::create([
-                'transaction_id' => $response['id'],
-                'payment_method' => 'paypal',
-                'user_id' => $cart->user_id,
-                'amount' => $amount,
-                'response' => json_encode($response)
-            ]);
-
-            $orderId = 'ORD'.rand(1111111,9999999);
-            $order = Order::create([
-                'order_number' => $orderId,
-                'transaction_id' => $transaction->id,
-                'user_id' => $cart->user_id,
-                'address_id' => $cart->address_id,
-                'total_amount' => $amount,
-                'status' => 'completed' 
-            ]);
-
-            foreach ($cart->detail as $key => $value) {
-                OrderDetail::create([
-                    'order_id' => $order->id,
-                    'product_id' => $value->product_id,
-                    'price' => $value->price,
-                    'qty' => $value->qty,
-                    'color' => $value->color,
-                    'size' => $value->size
+                $transaction = Transaction::create([
+                    'transaction_id' => $response['id'],
+                    'payment_method' => 'paypal',
+                    'user_id' => $cart->user_id,
+                    'amount' => $amount,
+                    'response' => json_encode($response)
                 ]);
-            }
 
-            Cart::where('id',$cartId)->delete();
-            CartDetail::where('cart_id',$cartId)->delete();
+                $orderId = 'ORD'.rand(1111111,9999999);
+                $order = Order::create([
+                    'order_number' => $orderId,
+                    'transaction_id' => $transaction->id,
+                    'user_id' => $cart->user_id,
+                    'address_id' => $cart->address_id,
+                    'total_amount' => $amount,
+                    'status' => 'completed' 
+                ]);
+
+                foreach ($cart->detail as $key => $value) {
+                    OrderDetail::create([
+                        'order_id' => $order->id,
+                        'product_id' => $value->product_id,
+                        'price' => $value->price,
+                        'qty' => $value->qty,
+                        'color' => $value->color,
+                        'size' => $value->size
+                    ]);
+                }
+
+                Cart::where('id',$cartId)->delete();
+                CartDetail::where('cart_id',$cartId)->delete();
+            }
 
             return redirect()
                 ->route('paypal.success')
